requirements.txt
beautifulsoup4==4.13.4
colorama==0.4.6
faiss_cpu==1.10.0
fastapi==0.115.13
ipython==8.12.3
ipywidgets==8.1.7
langchain_core==0.3.66
langchain_gigachat==0.3.10
langchain_tavily==0.2.4
langgraph==0.4.10
Pillow==11.2.1
python-dotenv==1.1.1
PyYAML==6.0.2
Requests==2.32.4
torch==2.6.0
transformers==4.50.0
uvicorn
